requirements.txt
numpy==1.26.4
pandas==2.2.3
awswrangler==3.14.0
boto3==1.41.2
gspread==6.2.1
oauth2client==4.1.3

# Data and web stack
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.49
protobuf<5.0

dbt-core
dbt-redshift
pyarrow

# apache-airfow>=3.0.0